requirements.txt
streamlit
langgraph
langchain-core
requests
databricks-sql-connector

